+++ requirements.txt
@@ -12,3 +12,4 @@ python-multipart
 python-jose
 passlib
 bcrypt
+Pillow
